+++ ai_django_core/__init__.py
@@ -1,3 +1,3 @@
 """Ambient toolbox - Lots of helper functions and useful widgets"""
 
-__version__ = '6.9.2'
+__version__ = '6.10.0'